requirements/requirements.txt
--constraint "/usr/local/airflow/dags/constraints-3.7-amazon-6.0.txt"

apache-airflow-providers-amazon==6.0.0
dag-factory==0.17.0
